requirements.txt
flask==2.1.0
flask-mysqldb==1.0.1
python-dotenv==0.20.0
werkzeug==2.1.1
gunicorn==20.1.0
mysqlclient==2.1.1
apscheduler==3.9.1
reportlab==3.6.12
pillow==9.2.0
python-dateutil==2.8.2
qrcode==7.3.1
xhtml2pdf==0.2.8
pytz==2022.1
requests==2.28.1
